% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_ggplot.R
\name{gg_text_size}
\alias{gg_text_size}
\title{A function to change the font size of the title, x-axis, y-axis, among other aspects.
Usually these all need to be adjusted for presentations.
The only merit to using this function over ggplot is that this function
organizes these similarly themed elements together, making them
easier to remember.  It also saves a little typing.}
\usage{
gg_text_size(p, main_title = 1, axis.title.x = 1, axis.title.y = 1,
  axis.text.x = 1, axis.text.y = 1, strip.text = NULL,
  legend.title = NULL, legend.text = NULL)
}
\arguments{
\item{p}{ggplot object}

\item{main_title}{changes title size through axis.title}

\item{axis.title.x}{changes relative size through axis.title.x}

\item{axis.title.y}{changes relative size through axis.title.y}

\item{axis.text.x}{changes relative size through axis.text.x.
Remember this function changes the size of text on the x-axis not the
main axis title (usually numbers at each tick mark).}

\item{axis.text.y}{changes relative size through axis.text.y.}

\item{strip.text}{changes relative size through strip.text}

\item{legend.title}{changes relative size through legend.title}

\item{legend.text}{changes relative size through legend.text}
}
\value{
ggplot object
}
\description{
A function to change the font size of the title, x-axis, y-axis, among other aspects.
Usually these all need to be adjusted for presentations.
The only merit to using this function over ggplot is that this function
organizes these similarly themed elements together, making them
easier to remember.  It also saves a little typing.
}
