% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_view.R
\name{hh}
\alias{hh}
\title{Wrapper for 'head()' to make for easier typing}
\usage{
hh(x)
}
\arguments{
\item{x}{object, usually a matrix or data.frame}
}
\value{
matrix or data.frame corner
}
\description{
Wrapper for 'head()' to make for easier typing
}
