% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_view.R
\name{ll}
\alias{ll}
\title{View the lower left corner of matrix or data.frame}
\usage{
ll(x, rowmax = 5, colmax = 5)
}
\arguments{
\item{x}{object, usually a matrix or data.frame}

\item{rowmax}{desired rows}

\item{colmax}{desired columns}
}
\value{
matrix or data.frame corner
}
\description{
View the lower left corner of matrix or data.frame
}
