% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myexpand.grid.R
\name{myexpand.grid}
\alias{myexpand.grid}
\title{a version of the expand.grid function that works more intuitively to me}
\usage{
myexpand.grid(..., stringsAsFactors = F)
}
\arguments{
\item{...}{varying factors that will appear in combinations
in the output}

\item{stringsAsFactors}{should characters be converted to factors}
}
\value{
Like expand.grid, "a data.frame containing a row
for every combination of supplied factors."  The main
difference is that the first factor varies the slowest
and, the last factor varies the slowest.  The names of
the factors are used as the names in the data.frame.
Also, stringsAsFactors is set to FALSE.
}
\description{
a version of the expand.grid function that works more intuitively to me
}
