% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_view.R
\name{ss}
\alias{ss}
\title{Wrapper for 'summary()' to make for easier typing}
\usage{
ss(x)
}
\arguments{
\item{x}{object}
}
\value{
matrix or data.frame corner
}
\description{
Wrapper for 'summary()' to make for easier typing
}
