% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_functions.R
\name{table_to_df}
\alias{table_to_df}
\title{Convert to results of table() to data.frame}
\usage{
table_to_df(x)
}
\arguments{
\item{table}{The output of table() function}
}
\value{
a data.frame
}
\description{
Convert to results of table() to data.frame
}
