% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_view.R
\name{uu}
\alias{uu}
\title{View the upper right corner of matrix or data.frame}
\usage{
uu(x, rowmax = 5, colmax = 5)
}
\arguments{
\item{x}{object, matrix or data.frame usually}

\item{rowmax}{desired rows}

\item{colmax}{desired columns}
}
\value{
matrix or data.frame corner
}
\description{
View the upper right corner of matrix or data.frame
}
